Finished file structure for folders and split off certain files
