+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.linting.pylintEnabled": false,
+    "python.pythonPath": "/Users/JonnyRiggs/.pyenv/versions/3.7.0/bin/python"
+}
