--- website/forms.py
@@ -1,16 +1,0 @@
-from django.contrib.auth.models import User
-from django import forms
-from website.models import Product
-
-class UserForm(forms.ModelForm):
-    password = forms.CharField(widget=forms.PasswordInput())
-
-    class Meta:
-        model = User
-        fields = ('username', 'email', 'password', 'first_name', 'last_name',)
-
-class ProductForm(forms.ModelForm):
-
-    class Meta:
-        model = Product
-        fields = ('title', 'description', 'price', 'quantity',)
