+++ website/forms/__init__.py
@@ -1,0 +1,2 @@
+from .product_form import ProductForm
+from .user_form import UserForm
