+++ website/forms/product_form.py
@@ -1,0 +1,10 @@
+from django.contrib.auth.models import User
+from django import forms
+
+from website.models import Product
+
+class ProductForm(forms.ModelForm):
+
+    class Meta:
+        model = Product
+        fields = ('title', 'description', 'price', 'quantity',)
