+++ website/forms/user_form.py
@@ -1,0 +1,9 @@
+from django.contrib.auth.models import User
+from django import forms
+
+class UserForm(forms.ModelForm):
+    password = forms.CharField(widget=forms.PasswordInput())
+
+    class Meta:
+        model = User
+        fields = ('username', 'email', 'password', 'first_name', 'last_name',)
