+++ website/models/__init__.py
@@ -1,0 +1,1 @@
+from .product_model import Product
