+++ website/views/__init__.py
@@ -1,0 +1,3 @@
+from .list_product_view import list_products
+from .sell_product_view import sell_product
+from .auth_view import login, login_required, login_user, logout
