+++ website/views/sell_product_view.py
@@ -1,0 +1,26 @@
+from website.forms import UserForm, ProductForm
+
+from website.models import Product
+
+
+
+
+def sell_product(request):
+    if request.method == 'GET':
+        product_form = ProductForm()
+        template_name = 'product/create.html'
+        return render(request, template_name, {'product_form': product_form})
+
+    elif request.method == 'POST':
+        form_data = request.POST
+
+        p = Product(
+            seller = request.user,
+            title = form_data['title'],
+            description = form_data['description'],
+            price = form_data['price'],
+            quantity = form_data['quantity'],
+        )
+        p.save()
+        template_name = 'product/success.html'
+        return render(request, template_name, {})
